Add feature idea
“The Manager's Path: A Guide for Tech Leaders Navigating Growth and
Change by” is a pretty big title and, unfortunately, Goodreads does
not seem to provide us with a shorter one. I was thinking we could
create a filter that, when a book gets matched with an item from the
filter, uses the title defined on it instead. For this book, for
example, it would use instead “The Manager's Path”.